% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{pf_get}
\alias{pf_get}
\title{Enumerates all files in a package}
\usage{
pf_get()
}
\description{
Wraps \code{\link[fs:dir_info]{fs::dir_info()}} and appends a column \code{"class"}.
Operates on the directory returned by \code{\link[usethis:proj_get]{usethis::proj_get()}}.
}
